% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citer.R
\name{citer_on_load}
\alias{citer_on_load}
\title{Set up the citation function on loading the package}
\usage{
citer_on_load(path = ".")
}
\arguments{
\item{path}{The path to the package directory. Defaults to the current
working directory.}
}
\description{
Set up the citation function on loading the package
}
