% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citer.R
\name{citer_write_environment}
\alias{citer_write_environment}
\title{Create/updates the \code{citer_environment.R} file in a package}
\usage{
citer_write_environment(path = ".", overwrite = TRUE)
}
\arguments{
\item{path}{The path to the package directory. Defaults to the current
working directory.}

\item{overwrite}{If \code{TRUE}, the existing \code{citer_environment.R} file will be
overwritten if it exists. If \code{FALSE}, an error will be raised if the file
already exists.}
}
\description{
Create/updates the \code{citer_environment.R} file in a package
}
\details{
This function creates or updates the \code{citer_environment.R} file in the
specified package directory. The file contains the citation information for
the package and is used to set up the citation function when the package is
loaded. The file is generated from the \code{citer} package's template and
includes the package name and a hash of the file content to ensure it is
up to date.

This function is called by the the \code{citer} setup functions.
}
\keyword{internal}
