% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_github_citations.R
\name{get_github_citations}
\alias{get_github_citations}
\title{Get information on packaged used by code on Github}
\usage{
get_github_citations(github_repos, verbose = TRUE, bibtex_file = NULL)
}
\arguments{
\item{github_repos}{character. A single repository or a vector of repositories.}

\item{verbose}{Logical. Default is TRUE. If TRUE displays messages on status.}

\item{bibtex_file.}{NULL or character. An optional file to which citations are written as a .bib file.}
}
\value{
Dataframe containing information on which packages are used by which R files.
}
\description{
Returns information needed to cite the GVS
}
\examples{
{
packages_used <- get_github_citations(github_repos = "https://github.com/Plant-Functional-Trait-Course/PFTC6_Norway",
                                      verbose = TRUE,
                                      bibtex_file = NULL)
}

}
