% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_package_citations.R
\name{get_package_citations}
\alias{get_package_citations}
\title{Get citation information for packages used}
\usage{
get_package_citations(packages_used_dataframe)
}
\arguments{
\item{packages_used_dataframe}{Output returned by \code{get_packages_used}}

\item{dependencies}{Logical. If TRUE, will additionally return information on the dependencies used by each R file.}
}
\value{
Dataframe containing information on which packages are used by which R files.
}
\description{
Returns information needed to cite R packages used in analyses
}
\examples{
{
packages_used_cites <- get_package_citations()
}

}
