% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_packages_used.R
\name{get_packages_used}
\alias{get_packages_used}
\title{Get information on packaged used}
\usage{
get_packages_used(..., detail = TRUE)
}
\arguments{
\item{...}{Additional arguments passed to internal functions.}
}
\value{
Dataframe containing bibtex-formatted citation information
}
\description{
Returns information needed to cite the GVS
}
\note{
This function provides citation information in bibtex format that can be used with reference manager software (e.g. Paperpile, Zotero). Please do remember to cite both the sources and the NSR, as the NSR couldn't exist without these sources!
}
\examples{
{
citation_info <- GVS_citations()
}

}
