% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citer.R
\name{citer_search_onLoad}
\alias{citer_search_onLoad}
\title{Function to search for \code{.onLoad} function in the package}
\usage{
citer_search_onLoad(path)
}
\description{
Function to search for \code{.onLoad} function in the package
}
