% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_packages_used.R
\name{get_packages_used}
\alias{get_packages_used}
\title{Get information on packaged used}
\usage{
get_packages_used(..., version_installed = FALSE, dependencies = FALSE)
}
\arguments{
\item{...}{R files and/or folders containing R files. If left blank, will default to the current working directory.}

\item{version_installed}{Logical. If TRUE, will return version information for installed packages.}

\item{dependencies}{Logical. If TRUE, will additionally return information on the dependencies used by each R file.}
}
\value{
Dataframe containing information on which packages are used by which R files.
}
\description{
Returns information needed for citation
}
\examples{
{
packages_used <- get_packages_used()
}

}
